server-chatbot/chatbot/requirements.txt
python-dotenv
openai
flask
flask-cors
